// src/Types.ts
export enum PrinterModel {
  BAMBU_X_SERIES = "BAMBU_X_SERIES",
  BAMBU_P_SERIES = "BAMBU_P_SERIES",
  PRUSA = "PRUSA",
}

export interface Printer {
  id?: string;
  name: string;
  ip: string;
  username?: string;
  accessCode?: string;
  modelType: PrinterModel;
  hasCamera?: boolean;
  serial?: string;
}

export interface PrinterStats {
  currentStatus: string;
  progressPercent: number;
  nozzleTemp: number;
  bedTemp: number;
}